New cli for clearer operation
A subtle cli is all well and good in that it allows you to do a variety
of things with few apparent changes to the command, but let's be real:
there are few **apparent** changes, but the changes are still there.
Subtlety, in this case, is more of a trap than it is any kind of
convenience.